Former-commit-id: 2679639f17fb1bb7a09bf7b1be45fc1e5dd12cb9

+++ app/Http/Controllers/ContractController.php
@@ -17,6 +17,9 @@ use App\Http\Requests\UploadContractFile;
 use App\Http\Resources\ContractResource;
 use App\Jobs\NotificationsJob;
 use App\Jobs\ProcessContractFile;
+use App\LocalCharCarrier;
+use App\LocalCharge;
+use App\LocalCharPort;
 use App\NewContractRequest;
 use App\NewContractRequestLcl;
 use App\Notifications\SlackNotification;
@@ -236,7 +239,7 @@ class ContractController extends Controller
             'carriers' => 'required',
         ]);
 
-        $status = $this->updateStatus($data['expire']);
+        $status = $this->updateStatus($contract, $data);
 
         $contract->update([
             'name' => $data['name'],
@@ -252,18 +255,22 @@ class ContractController extends Controller
         return new ContractResource($contract);
     }
 
-    public function updateStatus($data)
+    public function updateStatus($contract, $data)
     {
 
         $date = date('Y-m-d');
-        $expire = date('Y-m-d', strtotime($data));
-
-        if ($expire <= $date) {
-            $status = 'expired';
-        } else {
-            $status = 'publish';
+        $expire = date('Y-m-d', strtotime($data['expire']));
+        
+        if($contract->status != 'incomplete'){
+            if ($date <= $expire) {
+                $status = 'publish';
+            } else {
+                $status = 'expired';
+            }
+        }else{
+            $status = 'incomplete';
         }
-
+        
         return $status;
     }
 
@@ -363,7 +370,10 @@ class ContractController extends Controller
     public function removefile(Request $request, Contract $contract)
     {
         $media = $contract->getMedia('document')->where('id', $request->input('id'))->first();
-        $media->delete();
+        if(!empty($media) == 0){
+            $media->delete();
+        }
+        
 
         return response()->json(null, 204);
     }
@@ -561,7 +571,8 @@ class ContractController extends Controller
                     'status' => 'incomplete',
                     'type' => $type,
                     'gp_container_id' => 1,
-                    'code' => $request->reference,
+                    'code' => $code,
+                    'is_api' => 1,
                 ]);
                 break;
             case 'LCL':
@@ -573,7 +584,8 @@ class ContractController extends Controller
                     'expire' => $request->valid_until,
                     'status' => 'incomplete',
                     'type' => $type,
-                    'code' => $request->reference,
+                    'code' => $code,
+                    'is_api' => 1,
                 ]);
                 break;
         }
@@ -602,7 +614,7 @@ class ContractController extends Controller
                     'user_id' => Auth::user()->id,
                     'created' => date("Y-m-d H:i:s"),
                     'username_load' => 'Not assigned',
-                    'data' => '{"containers": [{"id": 1, "code": "20DV", "name": "20 DV"}, {"id": 2, "code": "40DV", "name": "40 DV"}, {"id": 3, "code": "40HC", "name": "40 HC"}, {"id": 4, "code": "45HC", "name": "45 HC"}, {"id": 5, "code": "40NOR", "name": "40 NOR"}], "group_containers": {"id": 1, "name": "DRY"}}',
+                    'data' => '{"containers": [{"id": 1, "code": "20DV", "name": "20 DV"}, {"id": 2, "code": "40DV", "name": "40 DV"}, {"id": 3, "code": "40HC", "name": "40 HC"}, {"id": 4, "code": "45HC", "name": "45 HC"}, {"id": 5, "code": "40NOR", "name": "40 NOR"}], "group_containers": {"id": 1, "name": "DRY"}, "contract":{"code":'.$contract->code.',"is_api":'.$contract->is_api.'}}',
                     'contract_id' => $contract->id,
                 ]);
                 break;
@@ -657,6 +669,18 @@ class ContractController extends Controller
         $contract = new Contract();
         $container = Container::get();
 
+        $data = $request->validate([
+            'referenceC' => 'required',
+            'group_containerC' => 'required',
+            'C20DV' => 'sometimes|required',
+            'C40DV' => 'sometimes|required',
+            'C40HC' => 'sometimes|required',
+            'C40NOR' => 'sometimes|required',
+            'C45HC' => 'sometimes|required',
+            'amountC' => 'sometimes|required',
+            'document' => 'required',
+        ]);
+
         $contract->company_user_id = Auth::user()->company_user_id;
         $contract->name = $request->referenceC;
         $validation = explode('/', $request->validityC);
@@ -665,7 +689,10 @@ class ContractController extends Controller
         $contract->expire = $validation[1];
         $contract->status = 'publish';
         $contract->gp_container_id = $request->group_containerC;
+        $contract->is_manual = 2;
         $contract->save();
+
+        $contract->ContractCarrierSyncSingle($request->carrierR);
 
         $rates = new Rate();
         $rates->origin_port = $request->origin_port;
@@ -678,7 +705,6 @@ class ContractController extends Controller
             $rates->fortyhc = $request->C40HC;
             $rates->fortynor = $request->C40NOR;
             $rates->fortyfive = $request->C45HC;
-
         } else {
 
             $rates->twuenty = 0;
@@ -701,14 +727,47 @@ class ContractController extends Controller
         $rates->contract()->associate($contract);
         $rates->save();
 
-        foreach ($request->input('document', []) as $file) {
-            $contract->addMedia(storage_path('tmp/uploads/' . $file))->toMediaCollection('document','contracts3');
+        // Surcharges
+
+        $calculation_type = $request->input('calculation');
+        $typeC = $request->input('type');
+        $currencyC = $request->input('currency');
+        $amountC = $request->input('amount');
+
+        if (count((array)$calculation_type) > 0) {
+            foreach ($calculation_type as $ct => $ctype) {
+
+                if (!empty($request->input('amount'))) {
+                    $localcharge = new LocalCharge();
+                    $localcharge->surcharge_id = $typeC[$ct];
+                    $localcharge->typedestiny_id = '3';
+                    $localcharge->calculationtype_id = $ctype;
+                    $localcharge->ammount = $amountC[$ct];
+                    $localcharge->currency_id = $currencyC[$ct];
+                    $localcharge->contract()->associate($contract);
+                    $localcharge->save();
+
+                    $detailcarrier = new LocalCharCarrier();
+                    $detailcarrier->carrier_id = $request->carrierR; //$request->input('localcarrier_id'.$contador.'.'.$c);
+                    $detailcarrier->localcharge()->associate($localcharge);
+                    $detailcarrier->save();
+
+                    $detailport = new LocalCharPort();
+                    $detailport->port_orig = $request->origin_port; // $request->input('port_origlocal'.$contador.'.'.$orig);
+                    $detailport->port_dest = $request->destination_port; //$request->input('port_destlocal'.$contador.'.'.$dest);
+                    $detailport->localcharge()->associate($localcharge);
+                    $detailport->save();
+                }
+            }
         }
 
+        foreach ($request->input('document', []) as $file) {
+            $contract->addMedia(storage_path('tmp/uploads/' . $file))->toMediaCollection('document', 'contracts3');
+        }
 
         return response()->json([
-            'data' => $rates->toJson(),
+            //'data' => $localcharge->toJson(),
+            'data' => 'Success',
         ]);
     }
-
 }
